+++ front/scripts/main/mixins/DiscussionRouteUpvoteMixin.ts
@@ -31,8 +31,7 @@ App.DiscussionRouteUpvoteMixin = Em.Mixin.create({
 
 			Em.$.ajax(<JQueryAjaxSettings>{
 				method: method,
-				url: 'https://' + M.prop('servicesDomain') +
-				'/discussion/' + Em.get(post, 'siteId') + '/votes/post/' + Em.get(post, 'id'),
+				url: M.getDiscussionServiceUrl(`/${Em.get(post,'siteId')}/votes/post/${Em.get(post,'id')}`),
 				dataType: 'json',
 				xhrFields: {
 					withCredentials: true,