+++ front/scripts/main/models/DiscussionForumModel.ts
@@ -19,8 +19,7 @@ App.DiscussionForumModel = Em.Object.extend(App.DiscussionErrorMixin, {
 	loadPage(pageNum: number) {
 		return new Em.RSVP.Promise((resolve: Function, reject: Function) => {
 			Em.$.ajax(<JQueryAjaxSettings>{
-				url: 'https://' + M.prop('servicesDomain') + '/discussion/' +
-					 this.wikiId + '/forums/' + this.forumId,
+				url: M.getDiscussionServiceUrl(`/${this.wikiId}/forums/${this.forumId}`),
 				data: {
 					page: pageNum
 				},
@@ -71,8 +70,7 @@ App.DiscussionForumModel.reopenClass({
 			}
 
 			Em.$.ajax(<JQueryAjaxSettings>{
-				url: 'https://' + M.prop('servicesDomain') +
-					 `/discussion/${wikiId}/forums/${forumId}`,
+				url: M.getDiscussionServiceUrl(`/${wikiId}/forums/${forumId}`),
 				data: requestData,
 				dataType: 'json',
 				xhrFields: {
@@ -84,14 +82,12 @@ App.DiscussionForumModel.reopenClass({
 						totalPosts = data.threadCount;
 
 					posts.forEach(function (post: any) {
-						var author: any;
 						if (post.hasOwnProperty('createdBy')) {
-							author = post.createdBy;
-							author.url = M.buildUrl({
+							post.createdBy.profileUrl = M.buildUrl({
 								namespace: 'User',
-								title: author.name
+								title: post.createdBy.name
 							});
-							contributors.push(author);
+							contributors.push(post.createdBy);
 						}
 					});