+++ front/scripts/mercury/modules/Trackers/Krux.ts
@@ -1,18 +1,18 @@
 'use strict';
 
-interface Window {
-	Krux: {
-		load?: (skinSiteId: string) => void;
-	};
-}
-
 module Mercury.Modules.Trackers {
 	export class Krux {
+
+		private kruxModule: {
+			load?: (skinSiteId: string) => void;
+		};
+
 		/**
+		 * @param {any} krux
 		 * @returns {void}
 		 */
-		constructor () {
-			window.Krux = window.Krux || [];
+		constructor (krux: any) {
+			this.kruxModule = krux || {};
 		}
 
 		/**
@@ -24,8 +24,8 @@ module Mercury.Modules.Trackers {
 		 * @returns {void}
 		 */
 		trackPageView (): void {
-			if (typeof window.Krux.load === 'function') {
-				window.Krux.load(M.prop('tracking.krux.mobileId'));
+			if (typeof this.kruxModule.load === 'function') {
+				this.kruxModule.load(M.prop('tracking.krux.mobileId'));
 			}
 		}
 	}